# src/Weaver.WebApp/packages/shared/src/api/models/option-type.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Weaver.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */

export type OptionType = typeof OptionType[keyof typeof OptionType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OptionType = {
  Number: 'Number',
  String: 'String',
  Boolean: 'Boolean',
  StringArray: 'StringArray',
  NumberArray: 'NumberArray',
} as const;