# src/Weaver.WebApp/packages/shared/src/api/models/problem-details.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Weaver.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */

export interface ProblemDetails {
  /** @nullable */
  type?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  status?: number | null;
  /** @nullable */
  detail?: string | null;
  /** @nullable */
  instance?: string | null;
}