# src/Weaver.WebApp/packages/shared/src/api/models/service-detail-model.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Weaver.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ServiceType } from './service-type';
import type { ServiceOption } from './service-option';

export interface ServiceDetailModel {
  id?: string;
  name: string;
  type: ServiceType;
  config?: ServiceOption[];
}