# src/Weaver.WebApp/packages/shared/src/api/models/service-list-item-model.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Weaver.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ServiceType } from './service-type';

export interface ServiceListItemModel {
  id?: string;
  name: string;
  type: ServiceType;
}