# src/Weaver.WebApp/packages/shared/src/api/models/service-option-model.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Weaver.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { OptionType } from './option-type';

export interface ServiceOptionModel {
  id: string;
  name: string;
  type: OptionType;
}