# src/Weaver.WebApp/packages/shared/src/api/models/service-option.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Weaver.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { OptionType } from './option-type';

export interface ServiceOption {
  name: string;
  type: OptionType;
  id?: number;
  uuid?: string;
}