# src/Weaver.WebApp/packages/shared/src/api/models/service-type.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Weaver.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */

export type ServiceType = typeof ServiceType[keyof typeof ServiceType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ServiceType = {
  Reference: 'Reference',
  Custom: 'Custom',
} as const;